Revive a UnitTest for Gamma Eos
This is derived from what used to be called UnitTest_Eos_3d a long time ago.
The name suggested now is UnitTest/Eos/gamma/3d/pm4dev.

This test currently fails in this branch, it demonstrates a failure
to return derived Eos quantities when using the Eos_vector implementation
of (at least) the Gamma Eos.