fix standalone use of macroProcessor.py
./macroProcessor.py can be run form the command line,
either with --macroDefs or without. In the latter case,
all regular .ini file in the current directory are loaded.

The --macroDefs argument can now be given more than once,
with the effect that macro definitions from several files
get accumulated.

(When macroProcessor.py is run from the command line, there is no
clear conflict resolution in case the same macro is defined in more
than one location; but expect a definition loaded later to override
an earlier one.)