macroProcessor: distinguish function-like macros
Function-like macros are macros with an argument list.  In other
words, function-like macros are macros that are expected to be
followed immediately (except perhaps for spaces) by a parenthesized
list, whereever they occur.

Up to now, the parsing of *all* occurrences of macros, whether
function-like or not, was gobbling up a parenthesized list if one was
following the macro reference. Here the parsing is changed, so that a
following parenthesized list is ignored while a non-functionlike macro
is expanded. This should avoid the need for tricks to "hide"
parentheses from the macro parser when parentheses occur as part of
the target (Fortran) language, particularly for argument lists.