A new functionality for `macroProcessor`
When requesting a variant for a specific `F90-mc` file,
it is necessary to change a subroutine naming and almost
every subroutine callers with `_VARIANT` suffix. For example,
the following source code:

   subroutine test_subroutine()
      use other_interface, ONLY: other_subroutineA, &
                                 other_subroutineB
      implicit none
      call other_subroutineA()
      call other_subroutineB()
   end subroutine test_subroutine

need to be processed to:

   subroutine test_subroutine_GPU()
      use other_interface, ONLY: other_subroutineA_GPU, &
                                 other_subroutineB_GPU
      implicit none
      call other_subroutineA_GPU()
      call other_subroutineB_GPU()
   end subroutine test_subroutine_GPU

when `GPU` variant is requested.

This can be achieved by using multiple macros, but
this commit introduces a simplified solution to this,
using a dedicated comment line at the top of a source code:

   !!VARIANTS(GPU): test_subroutine, &
   !!               other_subroutineA, &
   !!               other_subroutineB
   subroutine test_subroutine()
      use other_interface, ONLY: other_subroutineA, &
                                 other_subroutineB
      implicit none
      call other_subroutineA()
      call other_subroutineB()
   end subroutine test_subroutine

will result the expected behavior.