More tweaks for macro processing
Allow macro processing to proceed for .F90-mc files that
do not list all of the VARIANTS from the Config file line(s)
on
  !!VARIANTS(...)
lines in the file itself.

In such files, no changing of source language names inside the files
(subroutines names, etc.) will take place, for generated VARIANTS to
which this applies.
However, renaming of the source file can still take palce for such
a variant.